+++ domain/src/main/kotlin/com/sooum/domain/usecase/DeleteUserUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class DeleteUserUseCase @Inject constructor(
     private val repository: UserRepository
 ) {
-    suspend operator fun invoke(id: Long) {
+    suspend operator fun invoke(id: Int) {
         return repository.deleteUser(id)
     }
 }
